VRCSubs: Merge config keys instead of replacing them
This will fix a situation in which a referenced config key is NOT in the
config file, it will still use the default.

This is needed because if somebody is using the updater to update the
repo, Config.yml is NOT overwritten so it would result in KeyErrors
without this for new config vars.